Made adding pages very easy
