--- wp-content/themes/RoboTheme/divisions.php
@@ -1,12 +1,0 @@
-<?php
-	/*
-	Template Name: divisions
-	*/	
-?>
-<?php get_header(); ?>
-        <?php
-            
-        ?>
-	<?php get_footer(); ?>	
-    </body>
-</html>
