+++ wp-content/themes/RoboTheme/footer.php
@@ -1,0 +1,2 @@
+</body>
+</html>
